pulled out view functions from app
